add split component to view and logic
