+++ package.json
@@ -101,6 +101,7 @@
     "@toruslabs/customauth-react-native-sdk": "^1.0.1",
     "@toruslabs/torus-direct-web-sdk": "5.0.2",
     "@tradle/react-native-http": "^2.0.1",
+    "FormData": "^0.10.1",
     "abi-decoder": "^1.2.0",
     "abortcontroller-polyfill": "^1.7.3",
     "amplitude-js": "^8.8.0",
@@ -122,7 +123,6 @@
     "ethereumjs-tx": "^2.1.2",
     "eventemitter3": "^4.0.7",
     "events": "^1.1.1",
-    "FormData": "^0.10.1",
     "hdkey": "^1.1.2",
     "hex-to-rgba": "^2.0.1",
     "https-browserify": "0.0.1",
@@ -278,7 +278,6 @@
     "babel-plugin-named-asset-import": "0.3.4",
     "babel-plugin-react-native-web": "^0.12.0-rc.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
     "babel-preset-react-native": "^5.0.2",
     "bundlesize": "^0.18.0",
     "case-sensitive-paths-webpack-plugin": "^2.3.0",
